% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.imported.functions.r
\name{available.imported.functions}
\alias{available.imported.functions}
\alias{compmeans}
\alias{crosstab}
\alias{csv.get}
\alias{cut2}
\alias{ddply}
\alias{describe}
\alias{freq}
\alias{plotmeans}
\alias{scatterplot}
\alias{spss.get}
\alias{stata.get}
\alias{svyboxplot}
\alias{svyby}
\alias{svychisq}
\alias{svydesign}
\alias{svyglm}
\alias{svymean}
\alias{svyplot}
\alias{svytable}
\alias{wtd.boxplot}
\alias{wtd.chi.sq}
\alias{wtd.cor}
\alias{wtd.hist}
\alias{wtd.mean}
\alias{wtd.quantile}
\alias{wtd.t.test}
\alias{wtd.var}
\title{Imported Functions Made Available to Users}
\usage{
compmeans(...)

crosstab(...)

csv.get(...)

cut2(...)

ddply(...)

describe(...)

freq(...)

plotmeans(...)

scatterplot(...)

spss.get(...)

stata.get(...)

svyboxplot(...)

svyby(...)

svychisq(...)

svydesign(...)

svyglm(...)

svymean(...)

svyplot(...)

svytable(...)

wtd.boxplot(...)

wtd.chi.sq(...)

wtd.cor(...)

wtd.hist(...)

wtd.mean(...)

wtd.quantile(...)

wtd.t.test(...)

wtd.var(...)
}
\arguments{
\item{...}{See source package help file for details on function usage.}
}
\description{
The \code{\link[descr]{compmeans}} function is imported from descr package.  
See \code{\link[descr]{compmeans}} documentation for details.

The \code{\link[descr]{crosstab}} function is imported from the descr package.
See \code{\link[descr]{crosstab}} documentation for details.

The \code{\link[Hmisc]{csv.get}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{csv.get}} documentation for details.

The \code{\link[Hmisc]{cut2}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{cut2}} documentation for details.

The \code{\link[plyr]{ddply}} function is imported from the plyr package.  
See \code{\link[plyr]{ddply}} documentation for details.

The \code{\link[Hmisc]{describe}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{describe}} documentation for details.

The \code{\link[descr]{freq}} function is imported from the descr package. 
See \code{\link[descr]{freq}} documentation for details.

The \code{\link[gplots]{plotmeans}} function is imported from the gplots package.  
See \code{\link[gplots]{plotmeans}} documentation for details.

The \code{\link[car]{scatterplot}} function is imported from the car package.  
See \code{\link[car]{scatterplot}} documentation for details.

The \code{\link[Hmisc]{spss.get}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{spss.get}} documentation for details.

The \code{\link[Hmisc]{stata.get}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{stata.get}} documentation for details.

The \code{\link[survey]{svyboxplot}} function is imported from the survey package.  
See \code{\link[survey]{svyboxplot}} documentation for details.

The \code{\link[survey]{svyby}} function is imported from the survey package.  
See \code{\link[survey]{svyby}} documentation for details.

The \code{\link[survey]{svychisq}} function is imported from the survey package.  
See \code{\link[survey]{svychisq}} documentation for details.

The \code{\link[survey]{svydesign}} function is imported from the survey package.  
See \code{\link[survey]{svydesign}} documentation for details.

The \code{\link[survey]{svyglm}} function is imported from the survey package.  
See \code{\link[survey]{svyglm}} documentation for details.

The \code{\link[survey]{svymean}} function is imported from the survey package.  
See \code{\link[survey]{svymean}} documentation for details.

The \code{\link[survey]{svyplot}} function is imported from the survey package.  
See \code{\link[survey]{svyplot}} documentation for details.

The \code{\link[survey]{svytable}} function is imported from the survey package.  
See \code{\link[survey]{svytable}} documentation for details.

The \code{\link[ENmisc]{wtd.boxplot}} function is imported from the ENmisc package.  
See \code{\link[ENmisc]{wtd.boxplot}} documentation for details.

The \code{\link[weights]{wtd.chi.sq}} function is imported from the weights package.  
See \code{\link[weights]{wtd.chi.sq}} documentation for details.

The \code{\link[weights]{wtd.cor}} function is imported from the weights package.  
See \code{\link[weights]{wtd.cor}} documentation for details.

The \code{\link[weights]{wtd.hist}} function is imported from the weights package.  
See \code{\link[weights]{wtd.hist}} documentation for details.

The \code{\link[Hmisc]{wtd.mean}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{wtd.mean}} documentation for details.

The \code{\link[Hmisc]{wtd.quantile}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{wtd.quantile}} documentation for details.

The \code{\link[weights]{wtd.t.test}} function is imported from the weights package.  
See \code{\link[weights]{wtd.t.test}} documentation for details.

The \code{\link[Hmisc]{wtd.var}} function is imported from the Hmisc package.  
See \code{\link[Hmisc]{wtd.var}} documentation for details.
}

