% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI99.r
\name{CI99}
\alias{CI99}
\title{Calculates a 99 Percent Confidence Interval}
\usage{
CI99(m, se)
}
\arguments{
\item{m}{A single number equal to the mean value of a statistic.}

\item{se}{single number equal to the standard error of the statistic.}
}
\value{
Returns the 99% confidence interval as a vector of numeric values (the lower and upper bounds).
}
\description{
Prints the lower and upper boundaries of a 99 Percent confidence interval.  Also see the CI95 function.
}
\examples{
 
  CI99(10, 1)
  
  CI99(m=8, se=2)
}

