% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors.r
\name{Colors}
\alias{Colors}
\title{Displays R Colors}
\usage{
Colors()
}
\value{
No value returned
}
\description{
Produces plot of colors available in R. Numeric labels same for whole row.
}
\examples{
Colors()
}

