% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colPercents.r
\name{colPercents}
\alias{colPercents}
\title{Column Percentages}
\usage{
colPercents(tab, digits = 1)
}
\arguments{
\item{tab}{A table of values}

\item{digits}{Number of decimal places to display}
}
\value{
Table of column-percentaged values
}
\description{
Column percentages, based on function by John Fox
}

