% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compADPQ.r
\name{compADPQ}
\alias{compADPQ}
\title{Calculates a statistic for measure of association in cross-tabulation}
\usage{
compADPQ(x)
}
\arguments{
\item{x}{A matrix or data frame}
}
\value{
Returns a list of statistics
}
\description{
This function is called by \code{\link{tablesomersDC}} function.
}
