% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CramersV.r
\name{CramersV}
\alias{CramersV}
\title{Calculates Cramer's V}
\usage{
CramersV(chi, r, c, n)
}
\arguments{
\item{chi}{A number equal to the Chi-Squared statistic}

\item{r}{A number equal to the number of rows}

\item{c}{A number equal to the number of columns}

\item{n}{A number equal to the sample size}
}
\value{
The Cramer's V statistic, a number
}
\description{
Calculates Cramer's V, a measure of association to gauge the strength of the relationship between two nominal-level variables.  
  A score of 0 indicates no relationship; a score of 1 indicates a perfect relationship.
}
\examples{
  CramersV(84.18, 2, 2, 1315)
  CramersV(chi=84.18, r=2, c=2, n=1315)
}

