% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.svyglm.r
\name{fit.svyglm}
\alias{fit.svyglm}
\title{Calculates model fit statistic for a svyglm weighted regression model}
\usage{
fit.svyglm(svyglm)
}
\arguments{
\item{svyglm}{An object of type svyglm.  This object is the output of the svyglm functon.}
}
\value{
Returns classic model fit statistics, including R squared and Adjusted R Squared Statistic
}
\description{
model fit statistic for a svyglm weighted regression model
}

