% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqC.r
\name{freqC}
\alias{freqC}
\title{Generates table and figure describing distribution of variable values.}
\usage{
freqC(x, w)
}
\arguments{
\item{x}{a vector of variable values}

\item{w}{weights (optional)}
}
\value{
No value returned
}
\description{
Generates table and figure describing distribution of variable values. Based on freq() function in descr package.
}

