% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gss}
\alias{gss}
\title{GSS dataset for R Companion to Essentials of Political Analysis}
\format{A data frame with 1974 rows and 221 variables.}
\source{
Sources vary.  See book Appendix for further information.
}
\usage{
gss
}
\description{
The General Social Survey polls individuals about their attitudes and beliefs. This dataset is used to demonstrate application of R to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}

