% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gssD}
\alias{gssD}
\title{GSS design dataset for R Companion to Essentials of Political Analysis, Second Edition}
\format{
A design dataset version of the gss dataset used for certain weighted analyses.  The variables element has 1974 observations of 221 variables.
}
\source{
2012 General Social Survey.  See Appendix Table A.1 of printed textbook for further information
}
\usage{
gssD
}
\description{
The General Social Survey polls individuals about their attitudes and beliefs. This design dataset is used to demonstrate application of certain R functions to political analysis. See \code{\link{gss}} dataset documentation for variable names and descriptions.
}
\keyword{datasets}
