% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imeansC.r
\name{imeansC}
\alias{imeansC}
\title{Controlled Mean Comparison Table}
\usage{
imeansC(function1 = NULL, function2 = NULL, data)
}
\arguments{
\item{function1}{A function}

\item{function2}{A function}

\item{data}{Dataset}
}
\value{
No value returned
}
\description{
Produces a controlled mean comparison table
}

