% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.logit.r
\name{inverse.logit}
\alias{inverse.logit}
\title{Calculates predicted probability for a given logged odds value}
\usage{
inverse.logit(logged_odds)
}
\arguments{
\item{logged_odds}{a numeric value, or vector of numeric values}
}
\value{
Returns predicted probability corresponding the the logged odds value
}
\description{
This function calculates predicted probability for a given logged odds value
}

