% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplotC.r
\name{iplotC}
\alias{iplotC}
\title{Interaction plot}
\usage{
iplotC(function1 = NULL, function2 = NULL, data, function3 = NULL, ...)
}
\arguments{
\item{function1}{A function in the form ~depvar}

\item{function2}{A function. Here indepvar is typed before controlvar: ~indepvar + controlvar}

\item{data}{Design dataset (e.g. nesD, gssD, statesD, or worldD)}

\item{function3}{A function. Here indepvar is typed after controlvar: ~controlvar + indepvar}

\item{...}{Further arguments to be passed to brkdn plot}
}
\value{
No value returned
}
\description{
Interaction plot based on brkdn.plot function in plotrix package
}

