% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nes}
\alias{nes}
\title{NES dataset for R Companion to Essentials of Political Analysis}
\format{A data frame with 5916 rows and 399 variables.}
\source{
Sources vary.  See book Appendix for further information.
}
\usage{
nes
}
\description{
The American National Election Survey polls individuals about their political beliefs and behavior. This dataset is used to demonstrate application of R to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}

