% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orci.r
\name{orci}
\alias{orci}
\title{Translate logistic regression coefficients into odds ratios}
\usage{
orci(model, digits = 3)
}
\arguments{
\item{model}{An estimated logistic regression model}

\item{digits}{Number of digits after decimal to display}
}
\value{
Returns odds rations and confidence intervals in columns
}
\description{
Generates odds-ratios based on logistic regression model coefficients
}

