% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChar.r
\name{plotChar}
\alias{plotChar}
\title{Displays plotting characters available in R}
\usage{
plotChar()
}
\value{
No return
}
\description{
Displays plotting characters available in R. No values supplied to function.
}
\examples{
  plotChar()
}

