% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.testC.r
\name{prop.testC}
\alias{prop.testC}
\title{Difference of Proportions Test}
\usage{
prop.testC(y, x, w = NULL)
}
\arguments{
\item{y}{Dependent variable}

\item{x}{Independent variable}

\item{w}{Weights, optional}
}
\value{
No return
}
\description{
Difference of proportions test with optional weights
}

