% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickConsolePrint.r
\name{quickConsolePrint}
\alias{quickConsolePrint}
\title{Prints string to console quickly}
\usage{
quickConsolePrint(linetoprint)
}
\arguments{
\item{linetoprint}{Text to be printed to console}
}
\value{
No value returned
}
\description{
Prints string to console quickly, but still scolls for visual effect
}

