% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetTextContrastColor.r
\name{SetTextContrastColor}
\alias{SetTextContrastColor}
\title{Sets contrasting color for text}
\usage{
SetTextContrastColor(color)
}
\arguments{
\item{color}{A color}
}
\value{
Returns either "black" or "white"
}
\description{
Sets contrasting color for text
}

