% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slowConsolePrint.r
\name{slowConsolePrint}
\alias{slowConsolePrint}
\title{Prints string to console slowly}
\usage{
slowConsolePrint(linetoprint)
}
\arguments{
\item{linetoprint}{Text to be printed to console}
}
\value{
No value returned
}
\description{
Prints string to console slowly to make text more readable to the user
}

