% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somersD.r
\name{somersD}
\alias{somersD}
\title{Calculates Somer's D}
\usage{
somersD(formula, data)
}
\arguments{
\item{formula}{A formula}

\item{data}{Dataset to be sorted}
}
\value{
No return
}
\description{
This function makes use of the svytable function from the survey package
}

