% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortC.r
\name{sortC}
\alias{sortC}
\title{Returns case-level information in order}
\usage{
sortC(data, id, by, descending = TRUE)
}
\arguments{
\item{data}{Dataset to be sorted}

\item{id}{A variable in the dataset (data) that identfies individual cases, typically the name of states, countries, etc.}

\item{by}{Variable the dataset should be sorted by}

\item{descending}{Should the cases be sorted in descending order?  By default, set to TRUE}
}
\value{
A data frame
}
\description{
Returns case-level information in order
}

