% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svybyC.r
\name{svybyC}
\alias{svybyC}
\title{Computes summary statistics}
\usage{
svybyC(formula1 = NULL, formula2 = NULL, data, ...)
}
\arguments{
\item{formula1}{A formula}

\item{formula2}{A formula}

\item{data}{Dataset}

\item{...}{Further arguments to be passed}
}
\value{
A svyby-type object
}
\description{
Computes summary statistics
}

