% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablesomersDC.R
\name{tablesomersDC}
\alias{tablesomersDC}
\title{Produces summary statistics}
\usage{
tablesomersDC(x, dep = 2)
}
\arguments{
\item{x}{A svytable}

\item{dep}{which dimension stands for the dependent variable (1 = ROWS, 2 = COLS), default is 2.}
}
\value{
No return
}
\description{
This function is called by somersD function
}

