% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome.r
\name{welcome}
\alias{welcome}
\title{Function to Introduce Users to R Companion Environment}
\usage{
welcome()
}
\value{
No value returned
}
\description{
Welcomes users to Companion Package to Political Analysis and provides basic information about using Companion functions and datasets.
}

