% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{worldD}
\alias{worldD}
\title{World design dataset for R Companion to Essentials of Political Analysis}
\format{A survey.design object.  The variables element has 167 observations of 103 variables.}
\source{
Sources vary.  See book Appendix for further information.
}
\usage{
worldD
}
\description{
This is a design dataset with variables about countries in the world. This dataset is used to demonstrate application of R to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}

