% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.boxplot.r
\name{wtd.boxplot}
\alias{wtd.boxplot}
\title{Weighted Box Plot}
\usage{
wtd.boxplot(...)
}
\arguments{
\item{...}{Any arguments used will yield same warning message.}
}
\value{
Warning message only.
}
\description{
The wtd.boxplot function is no longer being used by the poliscidata package. Please use the svyboxplot function instead. 
See \code{\link[survey]{svyhist}} documentation for details; also see pp. 67-69 of The R Companion to Political Analysis, 2nd Edition.
}
