% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.median.r
\name{wtd.median}
\alias{wtd.median}
\title{Find Median of Variable}
\usage{
wtd.median(x, weights = NULL)
}
\arguments{
\item{x}{A variable}

\item{weights}{Sampling weights of variable (optional)}
}
\value{
Returns the median value of the variable
}
\description{
Takes in variable and finds median, works with sampling weights
}
\details{
Makes use of the wtd.quantile function, part of the Hmisc package
}

