% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.mode.r
\name{wtd.mode}
\alias{wtd.mode}
\title{Find Mode of Variable}
\usage{
wtd.mode(x, weights = NULL)
}
\arguments{
\item{x}{A variable}

\item{weights}{Sampling weights of variable (optional)}
}
\value{
Returns the modal value of the variable
}
\description{
Takes in variable and finds mode, works with sampling weights
}
\details{
Makes use of the freq function, part of the descr package
}

