% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.sd.r
\name{wtd.sd}
\alias{wtd.sd}
\title{Find Standard Deviation of Variable}
\usage{
wtd.sd(x, weights = NULL)
}
\arguments{
\item{x}{A variable (must be numeric), should be in dataset$var form.}

\item{weights}{Sampling weights of variable (optional), should be in dataset$weightvar form.}
}
\value{
Returns the standard deviation of the variable, a numeric value.
}
\description{
Takes in variable and calculates standard deviation, works with sampling weights.
}
\details{
Makes use of the freq function, part of the descr package.
}
\examples{
   library(poliscidata)
   
   wtd.sd(nes$polknow_combined, nes$wt)
}

