% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabC.r
\name{xtabC}
\alias{xtabC}
\title{Cross tabulation with optional weights}
\usage{
xtabC(function1, data)
}
\arguments{
\item{function1}{A function}

\item{data}{Dataset}
}
\value{
No return
}
\description{
Cross tabulation with optional weights
}

