% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtp.chi2.r
\name{xtp.chi2}
\alias{xtp.chi2}
\title{Cross tabulation with chi-squared statistic}
\usage{
xtp.chi2(data, y, x, w = NULL)
}
\arguments{
\item{data}{Dataset to be sorted}

\item{y}{Dependent variable}

\item{x}{Independent variable}

\item{w}{Weights (optional)}
}
\value{
Returns a data frame
}
\description{
This function makes use of the crosstab function from the descr package
}

