% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtp.r
\name{xtp}
\alias{xtp}
\title{Cross tabulation}
\usage{
xtp(data, y, x, w = NULL, ylab = NULL, xlab = NULL, main = NULL,
  dnn = NULL)
}
\arguments{
\item{data}{Dataset to be sorted}

\item{y}{Dependent variable}

\item{x}{Independent variable}

\item{w}{Weights (optional)}

\item{ylab}{Y-axis label (optional)}

\item{xlab}{X-axis label (optional)}

\item{main}{Main label for plot (optional)}

\item{dnn}{Optional, the names to be given to the dimensions in the result (the dimnames names)}
}
\value{
A data frame
}
\description{
Makes use of the crosstab function in the descr package.
}

