% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compADPQ.r
\name{compADPQ}
\alias{compADPQ}
\title{Calculates model fit statistic}
\usage{
compADPQ(x)
}
\arguments{
\item{x}{A matrix or data frame}
}
\value{
Returns a list
}
\description{
This function is called by the tablesomersDC function
}

