% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.logit.r
\name{inverse.logit}
\alias{inverse.logit}
\title{Calculates predicted probability for a given logged odds value}
\usage{
inverse.logit(logged_odds)
}
\arguments{
\item{logged_odds}{A numeric value, or vector of numeric values.}
}
\value{
Returns predicted probability corresponding the the logged odds value, a numeric value between 0 and 1.
}
\description{
This function calculates predicted probability for a given logged odds value, often useful for plotting or reporting predicted probabilities.
}
\examples{
   inverse.logit(0)
   
   inverse.logit(-5:5)
}
