% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logregR2.r
\name{logregR2}
\alias{logregR2}
\title{Logistic regression model statistics}
\usage{
logregR2(model, digits = 3)
}
\arguments{
\item{model}{An estimated logistic regression model}

\item{digits}{Number of digits to be displayed after decimal points}
}
\value{
Returns list of statistics about model
}
\description{
Logistic regression model statistics
}

