% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nesD}
\alias{nesD}
\title{NES design dataset for R Companion to Essentials of Political Analysis}
\format{A survey.design object.   The variables element has 5916 observations of 399 variables.}
\source{
Sources vary.  See book Appendix for further information.
}
\usage{
nesD
}
\description{
The American National Election Survey polls individuals about their political beliefs and behavior. This design dataset is used to demonstrate application of certain R functions to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}

