% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchisqC.r
\name{pchisqC}
\alias{pchisqC}
\title{Comparing Logistic Regression Models with Chi-Squared Test}
\usage{
pchisqC(reduced, full, digits = 3)
}
\arguments{
\item{reduced}{The reduced logisitic regression model as an object. This is model with fewer independent variables.}

\item{full}{The full logisitic regression model as an object. This is model with more independent variables.}

\item{digits}{The number of digits to display after decimal point, default is 3.}
}
\value{
The p-value of null hypothesis that full model no better than reduced model.
}
\description{
Comparing Logistic Regression Models with Chi-Squared Test
}

