% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmeansC.r
\name{plotmeansC}
\alias{plotmeansC}
\title{Plots mean comparison}
\usage{
plotmeansC(data, formula2 = NULL, formula3 = NULL, formula4, w = NULL,
  ...)
}
\arguments{
\item{data}{A dataset}

\item{formula2}{A formula}

\item{formula3}{A formula}

\item{formula4}{A formula}

\item{w}{Weights, optional}

\item{...}{Additonal arguments passed to plotmeans function}
}
\value{
No return
}
\description{
Plots mean comparison
}

