% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printC.r
\name{printC}
\alias{printC}
\title{Prints html-ready table to local working directory}
\usage{
printC(objx)
}
\arguments{
\item{objx}{A table or data frame}
}
\value{
No return
}
\description{
Prints html-ready table to local working directory
}

