% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somersD.r
\name{somersD}
\alias{somersD}
\title{Calculates Somer's D}
\usage{
somersD(formula, data)
}
\arguments{
\item{formula}{A formula in the form ~ indepvar + depvar}

\item{data}{A design dataset (e.g. gssD, nesD, statesD, or worldD)}
}
\value{
Return Somer's D measure of association statistic, a number between 0 and 1.
}
\description{
This function calculates the Somer's D measure of association statistic.  This function makes use of the svytable function from the survey package.
}
\examples{
   library(poliscidata)
   
   somersD(~ dem_educ3 + envjob_3, nesD)
}
