% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{states}
\alias{states}
\title{States dataset for R Companion to Essentials of Political Analysis}
\format{A data frame with 50 rows and 135 variables.}
\source{
Sources vary.  See book Appendix for further information.
}
\usage{
states
}
\description{
A dataset with variables about the 50 states. This dataset is used to demonstrate application of R to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}

