% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{statesD}
\alias{statesD}
\title{States design dataset for R Companion to Essentials of Political Analysis, Second Edition}
\format{A design dataset version of the states dataset used for certain weighted analyses.  The variables element has 50 observations of 135 variables.}
\source{
Sources vary.  See Appendix Table A.3 of printed textbook for further information
}
\usage{
statesD
}
\description{
This is a design dataset with variables about the 50 states. This design dataset is used to demonstrate application of certain R functions to political analysis. See \code{\link{states}} dataset documentation for variable names and descriptions.
}
\keyword{datasets}

