% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svychisqC.R
\name{svychisqC}
\alias{svychisqC}
\title{Chi-Squared Test with Weighted Data}
\usage{
svychisqC(formula, design)
}
\arguments{
\item{formula}{Formula expressing the relationship between variables}

\item{design}{Weighted design dataset}
}
\value{
No return
}
\description{
This function conducts a weighted chi-squared test
}

