% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{world}
\alias{world}
\title{World dataset for R Companion to Essentials of Political Analysis}
\format{A data frame with 167 rows and 103 variables.}
\source{
Sources vary.  See book Appendix for further information.
}
\usage{
world
}
\description{
A dataset with variables about countries in the world. This dataset is used to demonstrate application of R to political analysis. See book Appendix for variable names and descriptions.
}
\keyword{datasets}

