% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{worldD}
\alias{worldD}
\title{World design dataset for R Companion to Essentials of Political Analysis, Second Edition}
\format{
A design dataset version of the world dataset used for certain weighted analyses.  The variables element has 167 observations of 103 variables.
}
\source{
Sources vary.  See Appendix Table A.4 of printed textbook for further information
}
\usage{
worldD
}
\description{
This is a design dataset with variables about countries in the world. This dataset is used to demonstrate application of R to political analysis. See \code{\link{world}} dataset documentation for variable names and descriptions.
}
\keyword{datasets}
