% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.ttestC.R
\name{wtd.ttestC}
\alias{wtd.ttestC}
\title{Weighted T-Test}
\usage{
wtd.ttestC(f1, f2, data)
}
\arguments{
\item{f1}{A variable}

\item{f2}{Another variable}

\item{data}{Dataset}
}
\value{
No return
}
\description{
This function conducts a weighted t-test
}

